Fix fractional lootCount silently dropping nothing server-side
LootCount was truncated straight to [2]int at instance-state build time,
so a lootCount between 0 and 1 (meant to be the odds of a single drop)
always became [0,0] and never dropped anything. Carry it as [2]float64
through to a new resolveLootCount, which resolves the range to a value
and treats >= 1 as a truncated item count (unchanged behavior) and a
value in [0, 1) as the probability of awarding exactly one item.

+++ game-server/internal/instancestate/unit_state.go
@@ -173,8 +173,10 @@ type UnitState struct {
 	EquippedItems map[string]instanceconfig.EquippedItem
 	DamageStatKey string // "strength", "agility", or "" - see CharacterClass.DamageStatKey
 
-	LootTable map[string]int    // identifier → weight; nil means no loot
-	LootCount [2]int            // [min, max] items to award; both 1 when lootCount omitted
+	LootTable map[string]int // identifier → weight; nil means no loot
+	LootCount [2]float64     // [min, max]; both 1 when lootCount omitted. A resolved value
+	// >= 1 awards that many items (truncated); a value in [0, 1)
+	// is the probability of awarding exactly one item.
 	LootItems []PendingLootItem // rolled at death; nil until the unit dies
 
 	Status               UnitStatus
